Feat: Adding partial front End
