+++ src/app/cart/cartTables/cartTables.component.html
@@ -1,0 +1,20 @@
+<router-outlet>
+  <nb-card>
+    <nb-card-body>
+      <ng2-smart-table [settings]="settings" [source]="source" (deleteConfirm)="onDeleteConfirm($event)"></ng2-smart-table>
+
+    </nb-card-body>
+    <nb-card-footer>
+      <div class="row">
+        <div class="column" style="margin-left: auto; margin-right: auto;">
+          <button class="btn btn-hero-success main-btn"onclick="document.getElementById('priceTag').style.visibility='visible';">
+           <span>checkout your total</span>
+         </button>
+    </div>
+    <div class="column" style=" margin: auto; vertical-align: middle;">
+      <input id="priceTag" style="visibility:hidden;" value="{{price}}" readonly>
+    </div>
+  </div>
+    </nb-card-footer>
+  </nb-card>
+</router-outlet>
