--- src/app/cart/tables/tables.component.html
@@ -1,9 +1,0 @@
-<router-outlet>
-<nb-card>
-  <nb-card-body>
-    <ng2-smart-table [settings]="settings" [source]="source" (deleteConfirm)="onDeleteConfirm($event)" (editConfirm)="onSaveConfirm($event)"
- (createConfirm)="onCreateConfirm($event)" >
-    </ng2-smart-table>
-  </nb-card-body>
-</nb-card>
-</router-outlet>
